Refactored debug detection code to helper
